+++ services/web/server/src/simcore_service_webserver/tags/_handlers.py
@@ -1,24 +1,15 @@
 import functools
-import re
-from datetime import datetime
 
 from aiohttp import web
 from aiopg.sa.engine import Engine
-from models_library.api_schemas_webserver._base import InputSchema, OutputSchema
-from models_library.users import GroupID, UserID
-from pydantic import ConstrainedStr, Field, PositiveInt
 from servicelib.aiohttp.application_keys import APP_DB_ENGINE_KEY
 from servicelib.aiohttp.requests_validation import (
-    RequestParams,
-    StrictRequestParams,
     parse_request_body_as,
     parse_request_path_parameters_as,
 )
 from servicelib.aiohttp.typing_extension import Handler
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
-from servicelib.request_keys import RQT_USERID_KEY
 from simcore_postgres_database.utils_tags import (
-    TagDict,
     TagNotFoundError,
     TagOperationNotAllowedError,
     TagsRepo,
@@ -28,6 +19,15 @@ from .._meta import API_VTAG as VTAG
 from ..login.decorators import login_required
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
+from .schemas import (
+    TagCreate,
+    TagGet,
+    TagGroupCreate,
+    TagGroupPathParams,
+    TagPathParams,
+    TagRequestContext,
+    TagUpdate,
+)
 
 
 def _handle_tags_exceptions(handler: Handler):
@@ -45,71 +45,6 @@ def _handle_tags_exceptions(handler: Handler):
     return wrapper
 
 
-#
-# API components/schemas
-#
-
-
-class _RequestContext(RequestParams):
-    user_id: UserID = Field(..., alias=RQT_USERID_KEY)  # type: ignore[literal-required]
-
-
-class TagPathParams(StrictRequestParams):
-    tag_id: PositiveInt
-
-
-class ColorStr(ConstrainedStr):
-    regex = re.compile(r"^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$")
-
-
-class TagUpdate(InputSchema):
-    name: str | None = None
-    description: str | None = None
-    color: ColorStr | None = None
-
-
-class TagCreate(InputSchema):
-    name: str
-    description: str | None = None
-    color: ColorStr
-
-
-class TagAccessRights(OutputSchema):
-    # NOTE: analogous to GroupAccessRights
-    read: bool
-    write: bool
-    delete: bool
-
-
-class TagGet(OutputSchema):
-    id: PositiveInt
-    name: str
-    description: str | None = None
-    color: str
-
-    # analogous to UsersGroup
-    access_rights: TagAccessRights = Field(..., alias="accessRights")
-
-    @classmethod
-    def from_db(cls, tag: TagDict) -> "TagGet":
-        # NOTE: cls(access_rights=tag, **tag) would also work because of Config
-        return cls(
-            id=tag["id"],
-            name=tag["name"],
-            description=tag["description"],
-            color=tag["color"],
-            access_rights=TagAccessRights(  # type: ignore[call-arg]
-                read=tag["read"],
-                write=tag["write"],
-                delete=tag["delete"],
-            ),
-        )
-
-
-#
-# API handlers
-#
-
 routes = web.RouteTableDef()
 
 
@@ -119,7 +54,7 @@ routes = web.RouteTableDef()
 @_handle_tags_exceptions
 async def create_tag(request: web.Request):
     engine: Engine = request.app[APP_DB_ENGINE_KEY]
-    req_ctx = _RequestContext.parse_obj(request)
+    req_ctx = TagRequestContext.parse_obj(request)
     new_tag = await parse_request_body_as(TagCreate, request)
 
     repo = TagsRepo(user_id=req_ctx.user_id)
@@ -141,7 +76,7 @@ async def create_tag(request: web.Request):
 @_handle_tags_exceptions
 async def list_tags(request: web.Request):
     engine: Engine = request.app[APP_DB_ENGINE_KEY]
-    req_ctx = _RequestContext.parse_obj(request)
+    req_ctx = TagRequestContext.parse_obj(request)
 
     repo = TagsRepo(user_id=req_ctx.user_id)
     async with engine.acquire() as conn:
@@ -157,7 +92,7 @@ async def list_tags(request: web.Request):
 @_handle_tags_exceptions
 async def update_tag(request: web.Request):
     engine: Engine = request.app[APP_DB_ENGINE_KEY]
-    req_ctx = _RequestContext.parse_obj(request)
+    req_ctx = TagRequestContext.parse_obj(request)
     path_params = parse_request_path_parameters_as(TagPathParams, request)
     tag_updates = await parse_request_body_as(TagUpdate, request)
 
@@ -176,7 +111,7 @@ async def update_tag(request: web.Request):
 @_handle_tags_exceptions
 async def delete_tag(request: web.Request):
     engine: Engine = request.app[APP_DB_ENGINE_KEY]
-    req_ctx = _RequestContext.parse_obj(request)
+    req_ctx = TagRequestContext.parse_obj(request)
     path_params = parse_request_path_parameters_as(TagPathParams, request)
 
     repo = TagsRepo(user_id=req_ctx.user_id)
@@ -184,32 +119,6 @@ async def delete_tag(request: web.Request):
         await repo.delete(conn, tag_id=path_params.tag_id)
 
     raise web.HTTPNoContent(content_type=MIMETYPE_APPLICATION_JSON)
-
-
-#
-# Share API
-#
-
-
-class TagGroupPathParams(TagPathParams):
-    group_id: GroupID
-
-
-class TagGroupCreate(InputSchema):
-    read: bool
-    write: bool
-    delete: bool
-
-
-class TagGroupGet(OutputSchema):
-    gid: GroupID
-    # access
-    read: bool
-    write: bool
-    delete: bool
-    # timestamps
-    created: datetime
-    modified: datetime
 
 
 @routes.get(f"/{VTAG}/tags/{{tag_id}}/groups", name="list_tag_groups")
